rename: 요구사항에 따라 폴더명 변경
