+++ src/constant/gameRule.js
